modularized Sidebar Footer
