+++ lib/components/layouts/Sidebar/Footer/index.js
@@ -1,7 +1,4 @@
-import React, { forwardRef, useState } from "react";
-
-import classNames from "classnames";
-import { NavLink } from "react-router-dom";
+import React, { useState } from "react";
 
 import AppSwitcher from "./AppSwitcher";
 import NavIconWrapper from "../NavIconWrapper";
@@ -9,32 +6,7 @@ import ProfileSection from "../ProfileSection";
 import ProfileSectionTooltip from "../ProfileSection/Tooltip";
 import Tooltip from "../../../Tooltip";
 import Typography from "../../../Typography";
-
-const LINK_CLASS_NAMES =
-  "flex items-center no-underline select-none neeto-ui-sidebar__link";
-
-const Link = forwardRef(({ children, to, href, ...otherProps }, ref) => {
-  const isActive = location.pathname.includes(to);
-
-  return href ? (
-    <a
-      className={classNames(LINK_CLASS_NAMES, { active: isActive })}
-      href={href}
-      {...otherProps}
-    >
-      {children}
-    </a>
-  ) : (
-    <NavLink
-      className={LINK_CLASS_NAMES}
-      innerRef={ref}
-      to={to}
-      {...otherProps}
-    >
-      {children}
-    </NavLink>
-  );
-});
+import { Link } from "./Link";
 
 const Footer = ({
   appName,
